Commented out adminLogin method
Now, all users and admins can use the singular login page and the method will first check if the credentials inputted are those of an admin. If they are not, it will check if the credentials are from a user. Whichever returns first, will login the client under that criteria